requirement.txt
asgiref==3.9.1
Django==5.2.4
sqlparse==0.5.3
djangorestframework==3.14.0
django-cors-headers==3.14.0
celery==5.5.3
drf-yasg==1.21.10